Move leaf model state types into `leaf-model-types.ts`
