+++ src/components/thumbnail/thumbnail.test.tsx
@@ -3,7 +3,7 @@ import "@testing-library/jest-dom";
 import { render, screen } from "@testing-library/react";
 import { Thumbnail } from "./thumbnail";
 import { IContainer } from "../../hooks/use-model-state";
-import { IModelInputState, IModelOutputState } from "../app";
+import { IModelInputState, IModelOutputState } from "../../leaf-model-types";
 import { AlgaeEatersAmountType, EnvironmentType, FishAmountType, LeafDecompositionType, LeafEatersAmountType } from "../../utils/sim-utils";
 
 describe("Thumbnail component", () => {
